# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PoplarGrid API 服务器，提供项目、成员等操作功能",
        "title": "PoplarGrid 核心 API 服务器",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "paths": {
...
        "/users/{id}/applications": {
            "get": {
                "description": "根据分页参数获取用户的申请列表，支持分页和排序\\n当列表为空时，会返回 null 而不是空数组\\n如果 id 不是当前登录的用户 ID，则返回 400 错误",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "获取当前用户的申请（发出或收到）列表，支持分页",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认值为 1",
                        "name": "page_serial",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认值为 10",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "申请类型，0：发送的申请，1：收到的申请",
                        "name": "kind",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.ApplicationBasic"
                            }
                        }
                    },
                    "400": {
                        "description": "无效的请求参数",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/detail": {
            "get": {
                "description": "获取指定用户的详细信息，包括 ID、昵称、QQ 等",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "获取用户详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.UserDetail"
                        }
                    },
                    "400": {
                        "description": "无效的请求参数",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/invitations": {
            "get": {
                "description": "根据分页参数获取用户的邀请列表，支持分页和排序\\n当列表为空时，会返回 null 而不是空数组\\n如果 id 不是当前登录的用户 ID，则返回 400 错误",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invitation"
                ],
                "summary": "获取当前用户的邀请（发出或者收到）列表，支持分页",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认值为 1",
                        "name": "page_serial",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认值为 10",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "邀请类型，0：发送的邀请，1：收到的邀请",
                        "name": "kind",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.InvitationBasic"
                            }
                        }
                    },
                    "400": {
                        "description": "无效的请求参数",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/projects": {
            "get": {
                "description": "获取指定用户参与的所有项目列表，支持分页和复合查询",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "获取用户参与的项目列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认值为 1",
                        "name": "page_serial",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认值为 10",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "项目状态（位掩码），用于复合查询，默认不筛选查询",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "项目所属的作品集 ID，默认不筛选作品集",
                        "name": "workset_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.ProjectBasic"
                            }
                        }
                    },
                    "400": {
                        "description": "无效的请求参数",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/teams": {
            "get": {
                "description": "获取指定用户参与的所有汉化组列表，支持分页和排序",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "获取当前用户参与的汉化组列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认值为 1",
                        "name": "page_serial",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认值为 10",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.TeamBasic"
                            }
                        }
                    },
                    "400": {
                        "description": "无效的请求参数",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dtos.ApplicationBasic": {
            "type": "object",
            "properties": {
                "applicant_id": {
                    "description": "申请者 ID",
                    "type": "integer"
                },
                "id": {
                    "description": "申请 ID",
                    "type": "integer"
                },
                "nickname": {
                    "description": "申请者昵称",
                    "type": "string"
                },
                "project": {
                    "description": "所属项目的基本信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dtos.InnerProject"
                        }
                    ]
                },
                "role": {
                    "description": "申请的角色，使用位掩码表示",
                    "type": "integer"
                },
                "status": {
                    "description": "申请状态，0 pending, 1 accepted, 2 rejected",
                    "type": "integer"
                }
            }
        },
...
        "dtos.InnerProject": {
            "type": "object",
            "properties": {
                "project_id": {
                    "description": "所属项目 ID",
                    "type": "integer"
                },
                "project_title": {
                    "description": "项目标题",
                    "type": "string"
                },
                "status": {
                    "description": "项目状态，使用位",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dtos.ProjectOverallStatus"
                        }
                    ]
                },
                "workset_id": {
                    "description": "所属作品",
                    "type": "integer"
                },
                "workset_index": {
                    "description": "作品集内的序号",
                    "type": "integer"
                }
            }
        },
        "dtos.InvitationBasic": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "邀请 ID",
                    "type": "integer"
                },
                "invite_role": {
                    "description": "邀请的角色，使用位掩码表示",
                    "type": "integer"
                },
                "invitee_id": {
                    "description": "InviterNickname string `json:\"inviter_nickname\"` // 邀请者昵称",
                    "type": "integer"
                },
                "invitee_nickname": {
                    "description": "被邀请者昵称",
                    "type": "string"
                },
                "inviter_id": {
                    "description": "邀请者 ID",
                    "type": "integer"
                },
                "project": {
                    "description": "所属项目的基本信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dtos.InnerProject"
                        }
                    ]
                },
                "status": {
                    "description": "邀请状态，0 pending, 1 accepted, 2 rejected",
                    "type": "integer"
                }
            }
        },
...
        "dtos.MemberLabor": {
            "type": "object",
            "properties": {
                "joined_time": {
                    "description": "加入的",
                    "type": "string"
                },
                "labor_role": {
                    "description": "成员在项目中的角色，使用位掩码表示",
                    "type": "integer"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "user_id": {
                    "description": "用户 ID",
                    "type": "integer"
                }
            }
        },
...
        "dtos.ProjectBasic": {
            "type": "object",
            "properties": {
                "allow_auto_join": {
                    "description": "是否允许自动加入",
                    "type": "boolean"
                },
                "id": {
                    "description": "项目 ID",
                    "type": "integer"
                },
                "is_hidden": {
                    "description": "是否是隐藏项目",
                    "type": "boolean"
                },
                "is_published": {
                    "description": "是否已发布",
                    "type": "boolean"
                },
                "labors": {
                    "description": "补充的成员分工信息",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.MemberLabor"
                    }
                },
                "legacy_id": {
                    "description": "历史遗留序号",
                    "type": "integer"
                },
                "moetran_id": {
                    "description": "龙译 ID",
                    "type": "string"
                },
                "status": {
                    "description": "项目状态，使用位掩码表示",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dtos.ProjectOverallStatus"
                        }
                    ]
                },
                "title": {
                    "description": "项目名称",
                    "type": "string"
                },
                "workset_id": {
                    "description": "所属作品集 ID",
                    "type": "integer"
                },
                "workset_index": {
                    "description": "作品集内的序号",
                    "type": "integer"
                }
            }
        },
...
        "dtos.ProjectOverallStatus": {
            "type": "integer",
            "enum": [
                3,
                12,
                48,
                192,
                768
            ],
            "x-enum-comments": {
                "PROJECT_STATUS_LETTER_MASK": "二进制 48",
                "PROJECT_STATUS_PROOF_MASK": "二进制 12",
                "PROJECT_STATUS_PUBLISH_MASK": "二进制 768",
                "PROJECT_STATUS_REVIEW_MASK": "二进制 192",
                "PROJECT_STATUS_TRANSLATE_MASK": "二进制 3"
            },
            "x-enum-varnames": [
                "PROJECT_STATUS_TRANSLATE_MASK",
                "PROJECT_STATUS_PROOF_MASK",
                "PROJECT_STATUS_LETTER_MASK",
                "PROJECT_STATUS_REVIEW_MASK",
                "PROJECT_STATUS_PUBLISH_MASK"
            ]
        },
...
        "dtos.TeamBasic": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "团队 ID",
                    "type": "integer"
                },
                "name": {
                    "description": "团队名称",
                    "type": "string"
                }
            }
        },
...
        "dtos.UserDetail": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "id": {
                    "description": "用户 ID",
                    "type": "integer"
                },
                "last_active": {
                    "description": "上次活跃时间",
                    "type": "string"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "poplar_is_admin": {
                    "description": "是否是 panel 管理员",
                    "type": "boolean"
                },
                "qq_number": {
                    "description": "QQ 号",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "可选，在 service 层发生错误时提供详细信息",
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
    }
}